ref: dto renaming. param -> command 변경
